Moved the accessor to root level of the extension
